Share terminal audio writer telemetry lifecycle
